[#50] Finished merging staging branch
