+++ app/src/main/AndroidManifest.xml
@@ -23,14 +23,14 @@
         <activity android:name=".SendSmsInvitationActivity" />
         <activity android:name=".activityPage.HomePageActivity" />
         <!-- TODO The login page should load up first until we have a welcome page-->
-        <activity android:name=".activityPage.LoginPageActivity">
+        <activity android:name=".activityPage.LoginPageActivity"/>
+        <activity android:name=".activityPage.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity">
-        <activity android:name=".WelcomeActivity"
+        <activity android:name=".activityPage.WelcomeActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
         <activity android:name=".activityPage.SignUpPageActivity"/>
         <activity android:name=".fragments.UsersFragment" />
