+++ app/src/main/java/com/example/bigowlapp/activityPage/MainActivity.java
@@ -1,0 +1,20 @@
+package com.example.bigowlapp.activityPage;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.bigowlapp.R;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        Intent i = new Intent(MainActivity.this, WelcomeActivity.class);
+        startActivity(i);
+        finish();
+    }
+}
