+++ app/src/main/java/com/example/bigowlapp/activityPage/SignUpPageActivity.java
@@ -28,7 +28,7 @@ public class SignUpPageActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_sign_up);
         initialize();
     }
