+++ app/src/main/res/layout/welcome_activity.xml
@@ -5,4 +5,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/splash_screen"
-    tools:context=".WelcomeActivity" />
+    tools:context=".activityPage.WelcomeActivity" />
